Remove old stripe implementation
Since there is a new implementation in place, legacy gateway needs to be removed.

  - Rename new stripe to be just stripe

  - Reference to new stripe new stripe location

  - Update graphql schema

  - Remove legacy test suite
