+++ saleor/graphql/payment/schema.py
@@ -3,7 +3,7 @@ from graphql_jwt.decorators import permission_required
 
 from ..core.fields import PrefetchingConnectionField
 from .enums import PaymentGatewayEnum
-from .mutations import PaymentCapture, PaymentRefund, PaymentVoid, PaymentSecureConfirm
+from .mutations import PaymentCapture, PaymentRefund, PaymentSecureConfirm, PaymentVoid
 from .resolvers import resolve_payment_client_token, resolve_payments
 from .types import Payment
