+++ saleor/graphql/schema.graphql
@@ -1447,7 +1447,6 @@ enum GatewaysEnum {
   BRAINTREE
   RAZORPAY
   STRIPE
-  STRIPE_NEW
 }
 
 scalar GenericScalar
