+++ saleor/payment/gateways/stripe/__init__.py
@@ -1,25 +1,27 @@
-from typing import Dict
+from typing import List
 
 import stripe
 
 from ... import TransactionKind
-from ...interface import GatewayConfig, GatewayResponse, PaymentData
-from .forms import StripePaymentModalForm
+from ...interface import (
+    CreditCardInfo,
+    CustomerSource,
+    GatewayConfig,
+    GatewayResponse,
+    PaymentData,
+)
 from .utils import (
     get_amount_for_stripe,
     get_amount_from_stripe,
     get_currency_for_stripe,
     get_currency_from_stripe,
-    get_payment_billing_fullname,
     shipping_to_stripe_dict,
 )
 
 
 def get_client_token(**_):
-    """Not implemented for Stripe gateway currently.
-
-    The client token can be automatically generated by
-    Stripe's checkout.js or stripe.js.
+    """Not implemented for stripe gateway currently. The client token can be
+    generated by Stripe's checkout.js or stripe.js automatically.
     """
     return
 
@@ -27,114 +29,148 @@ def get_client_token(**_):
 def authorize(
     payment_information: PaymentData, config: GatewayConfig
 ) -> GatewayResponse:
-    client, error = _get_client(**config.connection_params), None
+    kind = TransactionKind.CAPTURE if config.auto_capture else TransactionKind.AUTH
+    client = _get_client(**config.connection_params)
+    capture_method = "automatic" if config.auto_capture else "manual"
+    currency = get_currency_for_stripe(payment_information.currency)
+    stripe_amount = get_amount_for_stripe(payment_information.amount, currency)
+    future_use = "off_session" if config.store_customer else "on_session"
+    customer_id = PaymentData.customer_id if payment_information.reuse_source else None
+    shipping = (
+        shipping_to_stripe_dict(payment_information.shipping)
+        if payment_information.shipping
+        else None
+    )
 
     try:
-        # Authorize with/without capture
-        response = _create_stripe_charge(
-            client=client,
-            payment_information=payment_information,
-            should_capture=config.auto_capture,
+        intent = client.PaymentIntent.create(
+            payment_method=payment_information.token,
+            amount=stripe_amount,
+            currency=currency,
+            confirmation_method="manual",
+            confirm=True,
+            capture_method=capture_method,
+            setup_future_usage=future_use,
+            customer=customer_id,
+            shipping=shipping,
         )
+        if config.store_customer and not customer_id:
+            customer = client.Customer.create(payment_method=intent.payment_method)
+            customer_id = customer.id
+        success = intent.status in ("succeeded", "requires_capture", "requires_action")
+        response = _success_response(
+            intent=intent, kind=kind, success=success, customer_id=customer_id
+        )
+        response = fill_card_details(intent, response)
     except stripe.error.StripeError as exc:
-        response = _get_error_response_from_exc(exc)
-        error = exc.user_message
-
-    kind = TransactionKind.CAPTURE if config.auto_capture else TransactionKind.AUTH
-
-    # Create response
-    return _create_response(
-        payment_information=payment_information,
-        kind=kind,
-        response=response,
-        error=error,
-    )
+        response = _error_response(kind=kind, exc=exc, payment_info=payment_information)
+    return response
 
 
 def capture(payment_information: PaymentData, config: GatewayConfig) -> GatewayResponse:
-    client, error = _get_client(**config.connection_params), None
-
-    # Get amount from argument or payment, and convert to stripe's amount
-    amount = payment_information.amount
-    stripe_amount = get_amount_for_stripe(amount, payment_information.currency)
-
+    client = _get_client(**config.connection_params)
+    intent = None
     try:
-        # Retrieve stripe charge and capture specific amount
-        stripe_charge = client.Charge.retrieve(payment_information.token)
-        response = stripe_charge.capture(amount=stripe_amount)
+        intent = client.PaymentIntent.retrieve(id=payment_information.token)
+        capture = intent.capture()
+        response = _success_response(
+            intent=intent,
+            kind=TransactionKind.CAPTURE,
+            success=capture.status in ("succeeded", "requires_action"),
+        )
+        response = fill_card_details(intent, response)
     except stripe.error.StripeError as exc:
-        response = _get_error_response_from_exc(exc)
-        error = exc.user_message
+        action_required = intent.status == "requires_action" if intent else False
+        response = _error_response(
+            kind=TransactionKind.CAPTURE,
+            exc=exc,
+            payment_info=payment_information,
+            action_required=action_required,
+        )
+    return response
 
-    # Create response
-    return _create_response(
-        payment_information=payment_information,
-        kind=TransactionKind.CAPTURE,
-        response=response,
-        error=error,
-    )
+
+def confirm(payment_information: PaymentData, config: GatewayConfig) -> GatewayResponse:
+    client = _get_client(**config.connection_params)
+    try:
+        intent = client.PaymentIntent(id=payment_information.token)
+        intent.confirm()
+        response = _success_response(
+            intent=intent,
+            kind=TransactionKind.CONFIRM,
+            success=intent.status == "succeeded",
+        )
+        response = fill_card_details(intent, response)
+    except stripe.error.StripeError as exc:
+        response = _error_response(
+            kind=TransactionKind.CONFIRM, exc=exc, payment_info=payment_information
+        )
+    return response
 
 
 def refund(payment_information: PaymentData, config: GatewayConfig) -> GatewayResponse:
-    client, error = _get_client(**config.connection_params), None
-
-    # Get amount from payment, and convert to stripe's amount
-    amount = payment_information.amount
-    stripe_amount = get_amount_for_stripe(amount, payment_information.currency)
-
+    client = _get_client(**config.connection_params)
+    currency = get_currency_for_stripe(payment_information.currency)
+    stripe_amount = get_amount_for_stripe(payment_information.amount, currency)
     try:
-        # Retrieve stripe charge and refund specific amount
-        stripe_charge = client.Charge.retrieve(payment_information.token)
-        response = client.Refund.create(charge=stripe_charge.id, amount=stripe_amount)
+        intent = client.PaymentIntent.retrieve(id=payment_information.token)
+        refund = intent["charges"]["data"][0].refund(amount=stripe_amount)
+        response = _success_response(
+            intent=intent,
+            kind=TransactionKind.REFUND,
+            success=refund.status == "succeeded",
+            amount=payment_information.amount,
+            currency=get_currency_from_stripe(refund.currency),
+        )
     except stripe.error.StripeError as exc:
-        response = _get_error_response_from_exc(exc)
-        error = exc.user_message
-
-    # Create response
-    return _create_response(
-        payment_information=payment_information,
-        kind=TransactionKind.REFUND,
-        response=response,
-        error=error,
-    )
+        response = _error_response(
+            kind=TransactionKind.REFUND, exc=exc, payment_info=payment_information
+        )
+    return response
 
 
 def void(payment_information: PaymentData, config: GatewayConfig) -> GatewayResponse:
-    client, error = _get_client(**config.connection_params), None
-
+    client = _get_client(**config.connection_params)
     try:
-        # Retrieve stripe charge and refund all
-        stripe_charge = client.Charge.retrieve(payment_information.token)
-        response = client.Refund.create(charge=stripe_charge.id)
+        intent = client.PaymentIntent.retrieve(id=payment_information.token)
+        refund = intent["charges"]["data"][0].refund()
+        response = _success_response(
+            intent=intent,
+            kind=TransactionKind.VOID,
+            currency=get_currency_from_stripe(refund.currency),
+            raw_response=refund,
+        )
     except stripe.error.StripeError as exc:
-        response = _get_error_response_from_exc(exc)
-        error = exc.user_message
+        response = _error_response(
+            kind=TransactionKind.VOID, exc=exc, payment_info=payment_information
+        )
+    return response
 
-    # Create response
-    return _create_response(
-        payment_information=payment_information,
-        kind=TransactionKind.VOID,
-        response=response,
-        error=error,
-    )
+
+def list_client_sources(
+    config: GatewayConfig, customer_id: str
+) -> List[CustomerSource]:
+    client = _get_client(**config.connection_params)
+    cards = client.PaymentMethod.list(customer=customer_id, type="card")["data"]
+    return [
+        CustomerSource(
+            id=c.id,
+            gateway="stripe",
+            credit_card_info=CreditCardInfo(
+                exp_year=c.card.exp_year,
+                exp_month=c.card.exp_month,
+                last_4=c.card.last4,
+                name_on_card=None,
+            ),
+        )
+        for c in cards
+    ]
 
 
 def process_payment(
     payment_information: PaymentData, config: GatewayConfig
 ) -> GatewayResponse:
-    # Stripe supports capture during authorize process. No need to run other steps.
-    # If 'config.auto_capture is True', it will also capture funds from the card.
     return authorize(payment_information, config)
-
-
-def create_form(
-    data: Dict, payment_information: PaymentData, connection_params: Dict
-) -> StripePaymentModalForm:
-    return StripePaymentModalForm(
-        data=data,
-        payment_information=payment_information,
-        gateway_params=connection_params,
-    )
 
 
 def _get_client(**connection_params):
@@ -142,86 +178,58 @@ def _get_client(**connection_params):
     return stripe
 
 
-def _get_stripe_charge_payload(
-    payment_information: PaymentData, should_capture: bool
-) -> Dict:
-    shipping = payment_information.shipping
-
-    # Get currency
-    currency = get_currency_for_stripe(payment_information.currency)
-
-    # Get appropriate amount for stripe
-    stripe_amount = get_amount_for_stripe(payment_information.amount, currency)
-
-    # Get billing name from payment
-    name = get_payment_billing_fullname(payment_information)
-
-    # Construct the charge payload from the data
-    charge_payload = {
-        "capture": should_capture,
-        "amount": stripe_amount,
-        "currency": currency,
-        "source": payment_information.token,
-        "description": name,
-    }
-
-    if shipping:
-        # Update shipping address to prevent fraud in Stripe
-        charge_payload["shipping"] = {
-            "name": name,
-            "address": shipping_to_stripe_dict(shipping),
-        }
-
-    return charge_payload
-
-
-def _create_stripe_charge(client, payment_information, should_capture: bool):
-    """Create a charge with specific amount, ignoring payment's total."""
-    charge_payload = _get_stripe_charge_payload(payment_information, should_capture)
-    return client.Charge.create(**charge_payload)
-
-
-def _create_response(
-    payment_information: PaymentData, kind: str, response: Dict, error: str
+def _error_response(
+    kind: TransactionKind,
+    exc: Exception,
+    payment_info: PaymentData,
+    action_required: bool = False,
 ) -> GatewayResponse:
-    # Get currency from response or payment
-    currency = get_currency_from_stripe(
-        response.get("currency", payment_information.currency)
-    )
-
-    amount = payment_information.amount
-    # Get amount from response or payment
-    if "amount" in response:
-        stripe_amount = response.get("amount")
-        if "amount_refunded" in response:
-            # This happens for partial catpure which will refund the left
-            # Then the actual amount should minus refunded amount
-            stripe_amount -= response.get("amount_refunded")
-        amount = get_amount_from_stripe(stripe_amount, currency)
-
-    # Get token from response or use provided one
-    token = response.get("id", payment_information.token)
-
-    # Check if the response's status is flagged as succeeded
-    is_success = response.get("status") == "succeeded"
     return GatewayResponse(
-        is_success=is_success,
-        action_required=False,
-        transaction_id=token,
+        is_success=False,
+        action_required=action_required,
+        transaction_id=payment_info.token,
+        amount=payment_info.amount,
+        currency=payment_info.currency,
+        error=exc.user_message,
         kind=kind,
-        amount=amount,
-        currency=currency,
-        error=error,
-        raw_response=response,
+        raw_response=exc.json_body or {},
+        customer_id=payment_info.customer_id,
     )
 
 
-def _get_error_response_from_exc(exc):
-    response = exc.json_body
+def _success_response(
+    intent: stripe.PaymentIntent,
+    kind: TransactionKind,
+    success: bool = True,
+    amount=None,
+    currency=None,
+    customer_id=None,
+    raw_response=None,
+):
+    currency = currency or get_currency_from_stripe(intent.currency)
+    return GatewayResponse(
+        is_success=success,
+        action_required=intent.status == "requires_action",
+        transaction_id=intent.id,
+        amount=amount or get_amount_from_stripe(intent.amount, currency),
+        currency=currency,
+        error=None,
+        kind=kind,
+        raw_response=raw_response or intent,
+        customer_id=customer_id,
+    )
 
-    # Some errors from stripe don't json_body as None
-    # such as stripe.error.InvalidRequestError
-    if response is None:
-        response = dict()
 
+def fill_card_details(intent: stripe.PaymentIntent, response: GatewayResponse):
+    charges = intent.charges["data"]
+    if not charges:
+        return response
+
+    card = intent.charges["data"][-1]["payment_method_details"]["card"]
+    response.card_info = CreditCardInfo(
+        last_4=card["last4"],
+        exp_year=card["exp_year"],
+        exp_month=card["exp_month"],
+        brand=card["brand"],
+    )
     return response
