+++ saleor/payment/gateways/stripe/utils.py
@@ -31,7 +31,6 @@ ZERO_DECIMAL_CURRENCIES = [
 
 def get_amount_for_stripe(amount, currency):
     """Get appropriate amount for stripe.
-
     Stripe is using currency's smallest unit such as cents for USD and
     stripe requires integer instead of decimal, so multiplying by 100
     and converting to integer is required. But for zero-decimal currencies,
@@ -61,7 +60,6 @@ def get_amount_from_stripe(amount, currency):
 
 def get_currency_for_stripe(currency):
     """Convert Saleor's currency format to Stripe's currency format.
-
     Stripe's currency is using lowercase while Saleor is using uppercase.
     """
     return currency.lower()
@@ -69,7 +67,6 @@ def get_currency_for_stripe(currency):
 
 def get_currency_from_stripe(currency):
     """Convert Stripe's currency format to Saleor's currency format.
-
     Stripe's currency is using lowercase while Saleor is using uppercase.
     """
     return currency.upper()
@@ -85,10 +82,14 @@ def get_payment_billing_fullname(payment_information: PaymentData) -> str:
 
 def shipping_to_stripe_dict(shipping: AddressData) -> Dict:
     return {
-        "line1": shipping.street_address_1,
-        "line2": shipping.street_address_2,
-        "city": shipping.city,
-        "state": shipping.country_area,
-        "postal_code": shipping.postal_code,
-        "country": dict(countries).get(shipping.country, ""),
+        "name": shipping.first_name + " " + shipping.last_name,
+        "phone": shipping.phone,
+        "address": {
+            "line1": shipping.street_address_1,
+            "line2": shipping.street_address_2,
+            "city": shipping.city,
+            "state": shipping.country_area,
+            "postal_code": shipping.postal_code,
+            "country": dict(countries).get(shipping.country, ""),
+        },
     }
