+++ tests/api/test_gift_card.py
@@ -1,6 +1,7 @@
 from datetime import date
 
 import graphene
+
 from tests.api.utils import get_graphql_content
 
 from .utils import assert_no_permission
